Improve strided kernel and implement globalcoalesce  and 1D Block tiling Kernel
Strided: ~6T FLOPS (Best Kernel)

+++ src/matrix.cpp
@@ -7,8 +7,8 @@
 using namespace std;
 
 void print_mat(float *matrix, size_t N) {
-  for (int i = 0; i < N; i++) {
-    for (int j = 0; j < N; j++)
+  for (size_t i = 0; i < N; i++) {
+    for (size_t j = 0; j < N; j++)
       printf("%8.3f ", matrix[i * N + j]);
     std::cout << "\n";
   }
@@ -18,8 +18,8 @@ void transpose(float *matrix, size_t N) {
 
   float tmp;
 
-  for (int i = 0; i < N; i++)
-    for (int j = 0; j < i; j++) {
+  for (size_t i = 0; i < N; i++)
+    for (size_t j = 0; j < i; j++) {
       tmp = matrix[i * N + j];
       matrix[i * N + j] = matrix[j * N + i];
       matrix[j * N + i] = tmp;
@@ -36,7 +36,7 @@ float *random_mat(size_t N, int hi) {
   float *mat = init_mat(N);
   float r;
 
-  for (int i = 0; i < N * N; ++i) {
+  for (size_t i = 0; i < N * N; ++i) {
     r = ((float)rand() / (RAND_MAX)) + 1;
     mat[i] = (rand() % hi) * r;
   }
@@ -47,17 +47,17 @@ float *random_mat(size_t N, int hi) {
 float *init_mat(size_t N) {
   float *mat = (float *)malloc(sizeof(float) * N * N);
 
-  for (int i = 0; i < N * N; ++i)
+  for (size_t i = 0; i < N * N; ++i)
     mat[i] = 0;
   return mat;
 }
 
 void sgemm_cpu(float *A, float *B, float *C, size_t N) {
 
-  for (int row = 0; row < N; row++)
-    for (int col = 0; col < N; col++) {
+  for (size_t row = 0; row < N; row++)
+    for (size_t col = 0; col < N; col++) {
       float sum = 0.0;
-      for (int k = 0; k < N; k++)
+      for (size_t k = 0; k < N; k++)
         sum += A[row * N + k] * B[k * N + col];
       C[row * N + col] = sum;
     }
@@ -65,7 +65,7 @@ void sgemm_cpu(float *A, float *B, float *C, size_t N) {
 
 bool same_matrix(float *A, float *B, size_t N, float tolerance) {
 
-  for (int i = 0; i < N * N; i++)
+  for (size_t i = 0; i < N * N; i++)
     if (fabs(A[i] - B[i]) > tolerance) {
       printf("Mismatch at (%zu,%zu), %f != %f\n", i / N, i % N, A[i], B[i]);
       return false;
@@ -73,8 +73,7 @@ bool same_matrix(float *A, float *B, size_t N, float tolerance) {
   return true;
 }
 
-float measure_gflops(std::function<void()> MatMul_kernel, size_t N,
-                     size_t iterations) {
+float measure_gflops(std::function<void()> MatMul_kernel, size_t N, size_t iterations) {
 
   float elapsed_time, total_elapsed_time = 0;
   cudaEvent_t beg, end;
@@ -83,7 +82,7 @@ float measure_gflops(std::function<void()> MatMul_kernel, size_t N,
 
   size_t flops = 2 * N * N * N;
 
-  for (int i = 0; i < iterations; ++i) {
+  for (size_t i = 0; i < iterations; ++i) {
     cudaEventRecord(beg);
     MatMul_kernel();
     cudaEventRecord(end);
